+++ lib/nretnil-fake-data.rb
@@ -1,4 +1,6 @@
 # Base Class
 require 'nretnil/fake-data/colors'
 require 'nretnil/fake-data/data'
+require 'nretnil/fake-data/names'
 require 'nretnil/fake-data/version'
+require 'nretnil/fake-data/words'